+++ src/converters/piral-ng/common.d.ts
@@ -1,3 +1,0 @@
-export * from './esm/SharedModule';
-export * from './esm/NgExtension';
-export * from './esm/ResourceUrlPipe';