+++ src/framework/piral-core/src/types/state.ts
@@ -297,7 +297,10 @@ export interface PiralActions extends PiralCustomActions {
    * @param path The name of the component.
    * @param component The component to use for rendering.
    */
-  setRoute<T = {}>(path: string, component: ComponentType<RouteComponentProps<T>>): void;
+  setRoute<T extends { [K in keyof T]?: string } = {}>(
+    path: string,
+    component: ComponentType<RouteComponentProps<T>>,
+  ): void;
   /**
    * Includes a new provider as a sub-provider to the current provider.
    * @param provider The provider to include.